hide serde behind a feature flag + add support for the treemap
