+++ Cargo.toml
@@ -19,9 +19,11 @@ license = "MIT OR Apache-2.0"
 bytemuck = "1.7.3"
 byteorder = "1.4.3"
 retain_mut = "=0.1.7"
+serde = { version = "1.0.139", optional = true }
 
 [features]
 simd = []
+serde = ["dep:serde"]
 
 [dev-dependencies]
 proptest = { version = "1.0.0" }
